Updates the development composition
We are no longer using ArangoDB, so it can be removed from the
composition.

Spark should not be using a local-caching DNS, so we should let
SPARK_PUBLIC_DNS be the default value. To simplify host lookups, we
should arrange a link with the Mongo container to allow hostname lookups.

Change-Id: I837e25746fc9ccb2707e9febbfe675f4e8d73f95