Changes to development composition to emulate Kubernetes deploy
This change adds the four core processing services (events, query,
schedule, execute) into the Docker composition. Each of the services
is added as a single "mixin" Docker Compose configuration. The Kafka
configuration now DEFAULTS to an "internal" kafka that DOES NOT work
from outside the internal composition network. An additional "values"
configuration has been added for when Kafka should be properly
exposed.

The correct combinations of configuration for particular
"compositions" are now done as compose* scripts in the dev/
directory. Notes within the scripts explain how each composition is
used.